[MSA-100] refactor: Company Repository 파일들 수정
