+++ company/src/main/java/com/i4u/company/domain/repository/CompanyJpaRepository.java
@@ -1,6 +1,6 @@
 package com.i4u.company.domain.repository;
 
-import com.i4u.company.domain.Company;
+import com.i4u.company.domain.entity.Company;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
