+++ company/src/main/java/com/i4u/company/domain/repository/CompanyRepository.java
@@ -1,12 +1,9 @@
 package com.i4u.company.domain.repository;
 
-import com.i4u.company.domain.Company;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
+import com.i4u.company.domain.entity.Company;
 
-import java.util.Optional;
 import java.util.UUID;
 
 @Repository
